EZP-25539: REST limitation value href parsing
Role limitations now have their own (internal) mediatype, and use their own parsers.

A RouteBased limitation parser is added, the gets the limitation value from a property
name and the provided href. Section uses that parser, but Subtree uses an override of
it, because of special requirements in value handling.